refactor: add memoized component as default export
